Init refactoring for Home-Assistant
