+++ example.py
@@ -1,6 +1,8 @@
 #!/usr/bin/python3
 import time, sys
 import logging
+from pyhomematic import EasyServer as pyhomematic
+
 logging.basicConfig(level=logging.INFO)
 
 DEVICE1 = 'address_of_rollershutter_device' # e.g. KEQ7654321
@@ -12,19 +14,18 @@ def systemcallback(src, *args):
     for arg in args:
         print(arg)
 
-import pyhomematic
 try:
     # Create a server that listens on 127.0.0.1:7080 and identifies itself as myserver.
     # Connect to Homegear at 127.0.0.1:2001
     # Automatically start everything. Without autostart, pyhomematic.start() can be called.
     # We add a systemcallback so we can see what else happens besides the regular events.
     pyhomematic.create_server(interface_id="myserver",
-                            local="127.0.0.1",
-                            localport=7080,
-                            remote="127.0.0.1",
-                            remoteport=2001,
-                            autostart=True,
-                            systemcallback=systemcallback)
+                              local="127.0.0.1",
+                              localport=7080,
+                              remote="127.0.0.1",
+                              remoteport=2001,
+                              autostart=True,
+                              systemcallback=systemcallback)
 except:
     sys.exit(1)
 
@@ -39,41 +40,41 @@ if pyhomematic.Server:
         sleepcounter += 1
         time.sleep(1)
     print(pyhomematic.devices)
-    
+
     # Get level of rollershutter from 0.0 to 1.0.
     print(pyhomematic.devices[DEVICE1].level)
-    
+
     # Set level of rollershutter to 50%.
     pyhomematic.devices[DEVICE1].level = 0.5
     time.sleep(10)
-    
+
     # Move rollershutter down.
     pyhomematic.devices[DEVICE1].move_down()
     time.sleep(10)
-    
+
     # Get level of rollershutter from 0.0 to 1.0 directly from channel.
     print(pyhomematic.devices_all[DEVICE1 + ':1'].getValue("LEVEL"))
-    
+
     # Check if doorcontact is open by querying the device.
     print(pyhomematic.devices[DEVICE2].is_open)
-    
+
     # Check if doorcontact is open or closed by querying the device-channel. True or False, depending on state.
     print(pyhomematic.devices_all[DEVICE2 + ':1'].getValue("STATE"))
 
     # Get Actual Temperature
     print(pyhomematic.devices[DEVICE3].actual_temperature)
-    
+
     # Get Set Temperature
     print(pyhomematic.devices[DEVICE3].set_temperature)
-    
+
     # Get Battery State
     print(pyhomematic.devices[DEVICE3].battery_state)
-    
+
     # Set an eventcallback for the doorcontact that should be called when events occur.
     pyhomematic.devices[DEVICE2].setEventCallback(eventcallback)
     time.sleep(10)
     # Now open / close doorcontact and watch the eventcallback being called.
-    
+
     # Stop the server thread so Python can exit properly.
     pyhomematic.stop()
 sys.exit(0)
